pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "AI browser agent that uses Claude's vision to perform guided web tasks"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "playwright>=1.49.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
browser-agent = "src.cli:main"
